+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ## NEW FEATURES
 
 - An OpenVeo icon is now displayed next to the plugin name in administration interface
+- Set player page title and header to the video title instead of the course name
 
 ## BUG FIXES